fix: unify the embedder env lock and correct the threshold's stated margin
Code-critic round on PR #3976, two MEDIUMs.

1. provider_tests.rs declared its OWN `tokio::sync::Mutex` ENV_LOCK, a
   different static from mod.rs's `std::sync::Mutex`. Two independent locks do
   not serialise against each other, so that file's TRUSTY_DEVICE /
   TRUSTY_EMBEDDER_MODEL mutations were never ordered against the tests in
   mod.rs — the same wrong-model race could still reappear (on macOS, or as
   soon as either #[ignore] there is lifted), and mod.rs's "one shared lock
   across all env-touching tests in this binary" claim was false crate-wide.

   Unified rather than caveated: the lock and its RAII guard move to a new
   `embedder::test_env` module at module scope, so `mod tests`, its
   `reference_accuracy_tests` child, and the sibling `provider_tests` all take
   the SAME static. The four tests in provider_tests.rs became synchronous to
   share the one sync lock — the two that need an async call build an explicit
   current-thread runtime, the same discipline this PR already applies to the
   reference-accuracy test, instead of forcing the whole tree onto an async
   lock that synchronous tests cannot take. Verified both #[ignore]d tests
   still pass under --ignored.

2. The threshold doc claimed 0.999 sits "~100x below the INT8 signature". That
   was wrong. As deficit-from-perfect ratios — which is what the threshold
   actually budgets — the 1e-3 budget is ~1000x larger than fp32's ~1e-6 drift
   (that half was right), but INT8 misses it by only ~10x with #3486's 0.9897,
   and ~4x with the 0.995703 reproduced for this issue. The comment now states
   both real ratios and draws the honest conclusion: a single order of
   magnitude is thin, which is exactly why the wrong-model discriminator is
   the model_name() assertion and not this number.

mod.rs's line-cap budget ratchets 559 -> 535 SLOC.

Closes #3711
